Move ThemeProps to shared
