+++ shared/src/components/FileMatchChildren.tsx
@@ -3,7 +3,7 @@ import { flatMap } from 'lodash'
 import * as React from 'react'
 import { Observable } from 'rxjs'
 import { Settings } from '../../../web/src/schema/settings.schema'
-import { ThemeProps } from '../../../web/src/theme'
+import { ThemeProps } from '../theme/theme'
 import { isSettingsValid, SettingsCascadeProps } from '../settings/settings'
 import { SymbolIcon } from '../symbols/SymbolIcon'
 import { toPositionOrRangeHash } from '../util/url'
