+++ web/src/enterprise/user/productSubscriptions/NewProductSubscriptionPageOrRedirectUser.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Redirect, RouteComponentProps } from 'react-router'
 import * as GQL from '../../../../../shared/src/graphql/schema'
-import { ThemeProps } from '../../../theme'
 import { UserSubscriptionsNewProductSubscriptionPage } from './UserSubscriptionsNewProductSubscriptionPage'
+import { ThemeProps } from '../../../../../shared/src/theme/theme'
 
 interface Props extends RouteComponentProps<{}>, ThemeProps {
     authenticatedUser: GQL.IUser | null
