+++ web/src/enterprise/user/productSubscriptions/UserSubscriptionsNewProductSubscriptionPage.tsx
@@ -11,10 +11,10 @@ import { asError, createAggregateError, ErrorLike } from '../../../../../shared/
 import { mutateGraphQL } from '../../../backend/graphql'
 import { HeroPage } from '../../../components/HeroPage'
 import { PageTitle } from '../../../components/PageTitle'
-import { ThemeProps } from '../../../theme'
 import { eventLogger } from '../../../tracking/eventLogger'
 import { BackToAllSubscriptionsLink } from './BackToAllSubscriptionsLink'
 import { ProductSubscriptionForm, ProductSubscriptionFormData } from './ProductSubscriptionForm'
+import { ThemeProps } from '../../../../../shared/src/theme/theme'
 
 interface Props extends RouteComponentProps<{}>, ThemeProps {
     /**
