+++ web/src/extensions/extension/RegistryExtensionContributionsPage.tsx
@@ -4,10 +4,10 @@ import { ContributableMenu } from '../../../../shared/src/api/protocol'
 import { ExtensionManifest } from '../../../../shared/src/schema/extensionSchema'
 import { asError, ErrorLike, isErrorLike } from '../../../../shared/src/util/errors'
 import { PageTitle } from '../../components/PageTitle'
-import { ThemeProps } from '../../theme'
 import { eventLogger } from '../../tracking/eventLogger'
 import { ExtensionAreaRouteContext } from './ExtensionArea'
 import { ExtensionNoManifestAlert } from './RegistryExtensionManifestPage'
+import { ThemeProps } from '../../../../shared/src/theme/theme'
 
 interface Props extends ExtensionAreaRouteContext, RouteComponentProps<{}>, ThemeProps {}
