+++ web/src/namespaces/NamespaceArea.tsx
@@ -1,8 +1,8 @@
 import { ExtensionsControllerProps } from '../../../shared/src/extensions/controller'
 import * as GQL from '../../../shared/src/graphql/schema'
-import { ThemeProps } from '../theme'
 import { RouteDescriptor } from '../util/contributions'
 import { PatternTypeProps } from '../search'
+import { ThemeProps } from '../../../shared/src/theme/theme'
 
 /**
  * Properties passed to all page components in the namespace area.
