+++ web/src/repo/RepoRevContainer.tsx
@@ -18,7 +18,7 @@ import { HeroPage } from '../components/HeroPage'
 import { ChromeExtensionToast } from '../marketing/BrowserExtensionToast'
 import { SurveyToast } from '../marketing/SurveyToast'
 import { IS_CHROME } from '../marketing/util'
-import { ThemeProps } from '../theme'
+import { ThemeProps } from '../../../shared/src/theme/theme'
 import { EventLoggerProps } from '../tracking/eventLogger'
 import { RouteDescriptor } from '../util/contributions'
 import { CopyLinkAction } from './actions/CopyLinkAction'
