+++ web/src/repo/blob/Blob.tsx
@@ -33,7 +33,7 @@ import {
 import { getHover } from '../../backend/features'
 import { WebHoverOverlay } from '../../components/shared'
 import { isDiscussionsEnabled } from '../../discussions'
-import { ThemeProps } from '../../theme'
+import { ThemeProps } from '../../../../shared/src/theme/theme'
 import { EventLoggerProps } from '../../tracking/eventLogger'
 import { DiscussionsGutterOverlay } from './discussions/DiscussionsGutterOverlay'
 import { LineDecorationAttachment } from './LineDecorationAttachment'
