+++ web/src/repo/compare/DiffHunk.tsx
@@ -8,7 +8,7 @@ import {
 import { LinkOrSpan } from '../../../../shared/src/components/LinkOrSpan'
 import * as GQL from '../../../../shared/src/graphql/schema'
 import { propertyIsDefined } from '../../../../shared/src/util/types'
-import { ThemeProps } from '../../theme'
+import { ThemeProps } from '../../../../shared/src/theme/theme'
 
 const DiffBoundary: React.FunctionComponent<{
     /** The "lines" property is set for end boundaries (only for start boundaries and between hunks). */
