+++ web/src/repo/compare/FileDiffNode.tsx
@@ -10,9 +10,9 @@ import { ExtensionsControllerProps } from '../../../../shared/src/extensions/con
 import * as GQL from '../../../../shared/src/graphql/schema'
 import { PlatformContextProps } from '../../../../shared/src/platform/context'
 import { FileSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../../shared/src/util/url'
-import { ThemeProps } from '../../theme'
 import { DiffStat } from './DiffStat'
 import { FileDiffHunks } from './FileDiffHunks'
+import { ThemeProps } from '../../../../shared/src/theme/theme'
 
 export interface FileDiffNodeProps extends PlatformContextProps, ExtensionsControllerProps, ThemeProps {
     node: GQL.IFileDiff
