+++ web/src/site-admin/SiteAdminAddExternalServicePage.tsx
@@ -10,7 +10,7 @@ import { renderMarkdown } from '../../../shared/src/util/markdown'
 import { mutateGraphQL } from '../backend/graphql'
 import { PageTitle } from '../components/PageTitle'
 import { refreshSiteFlags } from '../site/backend'
-import { ThemeProps } from '../theme'
+import { ThemeProps } from '../../../shared/src/theme/theme'
 import { ExternalServiceCard } from '../components/ExternalServiceCard'
 import { ExternalServiceVariant, getExternalService } from './externalServices'
 import { SiteAdminExternalServiceForm } from './SiteAdminExternalServiceForm'
